requirements.txt
Flask==2.0.1
flask-cors
opencv-python-headless==4.6.0.66
mediapipe==0.10.15
Werkzeug==2.0.3
matplotlib==3.4.3
numpy>=1.22,<1.24
gunicorn==20.1.0
requests
pydrive
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
fer
imageio
imageio-ffmpeg
Pillow
tensorflow-cpu==2.12.0
